Tweaks in mode functions and delay

+++ Coursework2.X/SounderDriver.h
@@ -6,7 +6,7 @@ extern "C" {
 #endif
 
 void InitialiseSounder(void); // Initialise the sounder.
-void BeepSounder(int); // Beep the sounder for amount of time.
+void BeepSounder(short); // Beep the sounder for amount of time.
